scroller/reelspace_api/serializers.py
from rest_framework import serializers
from reelspace.models import *



#serialize the User model
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'first_name', 'last_name', 'username', 'email', 'password')
        


