scroller/requirements.txt
Django==4.1
djangorestframework==3.14
django-cors-headers==3.10.0
tailwindcss
